test: Get HMAC tests passing
